Fix type conversion warning

+++ TDTChocolate/FoundationAdditions/NSTimeZone+TDTAdditions.m
@@ -15,12 +15,12 @@ static const NSInteger NumMinutesInHour = 60;
 @implementation NSTimeZone (TDTAdditions)
 
 - (NSString *)tdt_timeNumoffset {
-  NSInteger seconds = self.secondsFromGMT;
+  int seconds = (int)self.secondsFromGMT;
   NSString *sign = (seconds < 0) ? @"-" : @"+";
   seconds = abs(seconds);
 
   int minutes = (seconds / NumSecondsInMinute) % NumMinutesInHour;
-  int hours = (int)seconds / (NumSecondsInMinute * NumMinutesInHour);
+  int hours = seconds / (NumSecondsInMinute * NumMinutesInHour);
   
   return [NSString stringWithFormat:@"%@%02d:%02d", sign, hours, minutes];
 }
